tests/wpt/tests/css/css-viewport/zoom/widget.html

<!doctype html>
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1918838">
<link rel="help" href="https://drafts.csswg.org/css-viewport/#zoom-property">
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<input type=checkbox>
<input type=radio>
<script>
for (let i of document.querySelectorAll("input")) {
  test(function() {
    let unzoomed = i.getBoundingClientRect().width;
    i.style.zoom = "2";
    let zoomed = i.getBoundingClientRect().width;
    assert_not_equals(unzoomed, zoomed, "Zoom should be applied");
  }, i.type);
}
</script>
